Move grades to separate module and class
